Close the two code-scanning alerts open on the release branch
R1710 (`inconsistent-return-statements`, bitcoind.py). `get_vin_info_legacy()`
returns a triple from its `try` and ends its `except` on a call to
`raise_unresolved_prevout()`, which always raises -- but nothing said so to a
static reader, so the function read as returning `None` on the failure path,
i.e. a prevout resolution silently reported as successful. The docstring
already claimed "never returns"; it is now also a `NoReturn` annotation.

R0401 (`cyclic-import`, apiwatcher -> dbbuilder -> staterollback). The rollback
reached back into the watcher, through two local imports, for four names: the
replay cursor over `parsed_events` and the two event catalogs that decide when
`assets_info` is stale. The cursor -- the three INDEXED BY queries,
`get_last_block_parsed()`, `get_last_block_touched()` and the cached tip values
-- moves to a new leaf module `api/parsedevents.py`; the catalogs move to
`statetables.py`, beside the other rules the build and rollback paths already
share. Both sides now import what they need instead of importing each other,
and the local imports are gone.

`apiserver`, `healthz_server` and `wsgi` call `get_last_block_parsed()` through
`parsedevents` rather than re-exporting it from the watcher, so the module's
namespace does not outlive the move.

pylint is at 10.00/10 over the exact file set the scanner analyses; 1914 unit
tests pass (one pre-existing local failure deselected -- `chain_halt_test.py`
exercises the Rust deserializer against a stale wheel, and fails identically
without these changes).
