+++ release-notes/release-notes-v11.4.0.md
@@ -238,6 +238,7 @@ The differential tests above surfaced three ways in which a State DB maintained 
 
 - The rules for projecting a Ledger DB row into its State DB counterpart — decoding the compact `asset_index` / `address_id` / `(utxo_tx_hash, utxo_vout)` representations — now live in one place, `counterpartycore/lib/api/statetables.py`, shared by the build path (migrations `0004`, `0006`, `0014`) and the rollback path. Two independent copies of these rules would drift, and drift here is a silent divergence no API response would reveal.
 - The rules for deriving an event's address associations — `EVENTS_ADDRESS_FIELDS` and the two functions that write `address_events`, including the UTXO-owner alias lookup — moved out of `apiwatcher` into `counterpartycore/lib/api/addressevents.py`. Four places have to agree on them — the watcher, migration `0001`, the address-history backfill (#3503) and `parser/mempool.py` — and reaching back into the watcher for them created an `apiwatcher -> dbbuilder -> apiwatcher` import cycle, which the move breaks.
+- The State DB's replay cursor over `parsed_events` — the three `INDEXED BY` queries, `get_last_block_parsed()`, `get_last_block_touched()` and the cached tip values — moved out of `apiwatcher` into `counterpartycore/lib/api/parsedevents.py`, and the two event catalogs that decide when `assets_info` is stale (`ASSET_EVENTS`, `XCP_DESTROY_EVENTS`) into `statetables.py` beside the other rules the build and rollback paths share. The incremental rollback reads all of them, and reaching back into the watcher for them closed an `apiwatcher -> dbbuilder -> staterollback -> apiwatcher` import cycle.
 - New State DB migration `0016` adds the missing `block_index` indexes on `addresses`, `rps` and `rps_matches`.
 
 ## Bugfixes
